Added ability to give points to other players.
Points can now be given to other players by pressing the "Give Points"
button underneath the preview tab. A menu will popup asking for the
recipient and number of points to be given.

I also applied AddCSLuaFile to some of the new VGUI elements, I'm unsure
if this has any effect however. Better to be safe though eh.