Added support for ITEM.Slot
Allows an item to have a variable set for ITEM.Slot.
If an Item is equipped with the same slot, it well automatically be unequipped.
I added this because I needed it for something I was doing in pointshop and was really confused as to why it wasn't a thing already. Might aswell add it here.