Optimize item loading with a 50ms delay
With an approximation of over 150 items in the pointshop, the Garry's Mod game client will start hitching, because too many 'DPointShopItem' objects are being created at once. This optimization creates a 50ms delay, when each 'DPointShopItem' object is created. It allows the rest of the GUI to continue loading before all the items have completed. Hitching is reduced by approximately 80% and loading times are also reduced. This should theoretically triple the number of possible items, without an extended freeze period.